ai_engine/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bharatledger-ai-engine"
version = "0.1.0"
description = "BharatLedger AI invoice processing engine (OCR + LLM + GST)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "pdf2image>=1.16.0",
    "httpx>=0.26.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_engine*"]
